feat: Add enemies with random speeds and positions
Add enemies at random vertical positions and with random speed. Vertical positions are selected randomly from an array of 3 available elements.
There can be 4 possible speeds for the enemies and they are selected among 4 possible values from an array of 4 items.
x position is fixed and given so each enemy starts outside the canvas on the left.
Enemies movement is only possible from left to right.