feat: Add mouse and touch controls
Add mouse and touch controls to move the player. The player reacts to a click of the mouse or to a touch inside the game field and moves one step in whatever direction corresponds to the farthest coordinate of the point from the player image center (after removing the extra space that surrounds the image player inside the sprite).