fix: Fix mouse and touch actions at game start
Fix mouse and touch actions to react appropriately to a click or to a touch at game start.
Previously there were 2 event listeners for touch events and 2 for click events. This way when the user interacted with the game to start it, the user action triggered also another function that updated the player position showing it one tile above to where it should have been.
The easiest fix for that was to use only one event listener for mouse interactions and one for touch interactions. The listener function would therefore perform its action based on the stage of the game at that time, after reading the 'active' flag variable value.
Note that this behavior only appeared in non-desktop devices.