fix: Correct coordinates and variable names
Corrected coordinates when calculating the position of a mouse click or a touch by the user in the canvas to move the player.
The position of the mouse or touch press is now calculated against the center of the effective player image, while previously it was calculated taking the starting x and y positions of the sprite. This led to the player moving to a wrong tile when the click was very close.
Inside the 'touchDown' function the variables startTouchX and startTouchY were changed with touchX and touchY. Both these ones and mouseX, mouse Y were set as global.
The coordinates of the text to start the game were assigned to global variables, too.