feat: Make canvas responsive
Make the game scale to fit on different viewports.
The game was originally set at a fix size, but portability wasn't addressed. The game was big enough to be displayed clearly on a desktop, whereas on mobiles it was either too small or too big, with some areas hidden from the user, who needed to move a scrollbar.
So it was added the function 'calculateCanvasSize()' to set the canvas width and height to the effective viewport width and height. This function lets the game scene and characters be drawn to the canvas according to its effective scale factor, that is computed comparing the viewport dimensions with the maximum dimensions given to the canvas for desktop. The global variable 'scaleFactor' is used to scale appropriately every result that affects canvas coordinates (positions of tiles, enemies and player).
Two more event listeners were added to register for a resize event (on desktop) or a change in the orientation on mobiles that would change the size of the viewport.