feat: Make player collect gems and increment score
Add the possibility to collect gems for the player as it touches them. Only 3 gems can be displayed in the game and each one of them adds different points to the player total score.
The remaining gems to be displayed are counted by the variable 'remainingGems' in engine.js. That variable is used to render the gems and to replace the collected ones when it is positive.
As the player collects the gems, its score increases.
It's been decided to display one gem for each color and to add different points to them, to make things more interesting.