feat: Add rocks as obstacles to the game
Add rocks to the game field and make them behave like obstacles for the player.
To detect an obstacle, a function has been added to the player class that checks if the next move will place the player in the same spot of a rock. This function will perform its check when the user presses an arrow key to move the player.