feat: Let player collect heart and increse lives
Make the heart collectable and add lives as player property.
The player starts the game with 1 life; he can increment its lives by 1 unit as he collects a heart. The maximum number of lives that is possible to have is 2.
When an enemy collides the player, his lives decrease by 1 unit, while the enemy explodes.