fix: Fix gems rendering
Changed code to avoid gem overlap in the rendering process. Prior to this change it was possible that two consecutive gems were displayed in the same position, due to the low number of possible values for its x and y coordinates. When the player collected a gem, it was possible that he took two of them without even realizing it; the only clue was the update in his score.
A variable was introduced to take the value of the last gem horizontal coordinate that would now be compared with the selected one, to make sure that only a different coordinate would be accepted by the program.
Also it was inserted a control to make sure that gems could be collected only in world 2.