fix: Fix obstacles detection on small viewports
Updated function hasNotPathFree of Player prototype in app.js to let the player detect obstacles (rocks) on small viewports.
When the viewport was resized, the function compared two floating numbers with a very long decimal part and the problem was likely in a different decimal digit between the player and the obstacle coordinates. This behavior made the rock invisible to the player path and no obstacle was detected.
The function 'Math.floor' was added to round all the coordinate numbers to their integer part.