fix: Update image effective coordinates at resize
Added code to update images offset and effective x and y coordinates when viewport is resized.
Images were rendered correctly on the game scene, but the real coordinates of what was really visible inside the sprites of every entity in the game kept the value that was assigned to it when it was created.
So, for example, the key image has an effective height of (70 * 0.6)px inside its sprite that has a total height of (171 * 0.6)px if the game is loaded with a scale factor of 1 (that means at its full size). Before this update, if the scale factor changed because of a rotation of the device or a zoom, the scalefactor variable did not change. This way the visual impression was that sometimes, when the viewport was reduced, the player collected the heart or the key without entering the correct tile. Also there were problems in collecting gems in this same conditions.
The meta tag in index.html was corrected, too, to display the game at full size on mobile devices preventing pinch to zoom and handling rotation properly.