fix: Render correct screen at game over
Fixed a problem with the screen that showed the game status after the session had finished.
The flag variable that controlled if the player had won the game, once set to true after the first win, was not set back to false, which made the game show always the message 'game won' afterwards, even when it was not true.